使用ButterKnife框架代替传统的view查找,修改部分布局等

+++ trainingcamp/src/main/java/com/iyuba/trainingcamp/adapter/SentenceAdapter.java
@@ -62,7 +62,7 @@ public class SentenceAdapter extends RecyclerView.Adapter<SentenceAdapter.ViewHo
         }
         Typeface tf = Typeface.createFromAsset(mContext.getAssets(),"font/DINMedium_0.ttf");
         holder.score.setTypeface(tf);
-        holder.score.setText(String.valueOf(points));
+        holder.score.setText(String.valueOf((int)points));
         holder.SentenceView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
